% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{\%==\%}
\alias{\%==\%}
\title{Special Relational Operator}
\usage{
lhs \%==\% rhs
}
\arguments{
\item{lhs}{left hand value}

\item{rhs}{right hand value}
}
\value{
TRUE if values are equal, FALSE otherwise (treats \code{NA} as a value)
}
\description{
Special Relational Operator
}
