% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{podci_ado_prnt}
\alias{podci_ado_prnt}
\title{PODCI Adolescent Parent-Report Example Data}
\format{
A 3 x 86 \link[dplyr:reexports]{dplyr::tibble}
\enumerate{
\item During last week, easy/hard to: Lift heavy books?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item During last week, easy/hard to: Pour a half gallon of milk?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item During last week, easy/hard to: Open a jar that has been opened before?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item During last week, easy/hard to: Use a fork and spoon?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item During last week, easy/hard to: Comb his/her hair?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item During last week, easy/hard to: Button buttons?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item During last week, easy/hard to: Put on his/her coat?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item During last week, easy/hard to: Write with a pencil?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item Over the last 12 months, how often did child miss school because of health?
\itemize{
\item 1, Rarely | 2, Once a month | 3, Two or three times a month | 4, Once a week | 5, More than once a week | 6, Does not attend school
}
\item During last week, how happy with: looks?
\itemize{
\item 1, Very happy | 2, Somewhat happy | 3, Not sure | 4, Somewhat unhappy | 5, Very unhappy | 6, Child is too young
}
\item During last week, how happy with: body?
\itemize{
\item 1, Very happy | 2, Somewhat happy | 3, Not sure | 4, Somewhat unhappy | 5, Very unhappy | 6, Child is too young
}
\item During last week, how happy with: clothes or shoes can wear?
\itemize{
\item 1, Very happy | 2, Somewhat happy | 3, Not sure | 4, Somewhat unhappy | 5, Very unhappy | 6, Child is too young
}
\item During last week, how happy with: ability to do the same things friends do?
\itemize{
\item 1, Very happy | 2, Somewhat happy | 3, Not sure | 4, Somewhat unhappy | 5, Very unhappy | 6, Child is too young
}
\item During last week, how happy with: health in general?
\itemize{
\item 1, Very happy | 2, Somewhat happy | 3, Not sure | 4, Somewhat unhappy | 5, Very unhappy | 6, Child is too young
}
\item During last week, how often: feel sick and tired?
\itemize{
\item 1, Most of the time | 2, Some of the time | 3, A little of the time | 4, None of the time
}
\item During last week, how often: full of pep and energy?
\itemize{
\item 1, Most of the time | 2, Some of the time | 3, A little of the time | 4, None of the time
}
\item During last week, how often: pain or discomfort interfere with activities?
\itemize{
\item 1, Most of the time | 2, Some of the time | 3, A little of the time | 4, None of the time
}
\item During last week, easy/hard to: Run short distances?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item During last week, easy/hard to: Bicycle or tricycle?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item During last week, easy/hard to: Climb three flights of stairs?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item During last week, easy/hard to: Climb one flight of stairs?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item During last week, easy/hard to: Walk more than a mile?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item During last week, easy/hard to: Walk three blocks?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item During last week, easy/hard to: Walk one block?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item During last week, easy/hard to: Get on and off a bus?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too young for this activity
}
\item How often need help from another person for walking and climbing?
\itemize{
\item 1, Never | 2, Sometimes | 3, About half the time | 4, Often | 5, All the time
}
\item How often use assistive devices for walking and climbing?
\itemize{
\item 1, Never | 2, Sometimes | 3, About half the time | 4, Often | 5, All the time
}
\item During last week, easy/hard to:  Stand while washing hands and face at a sink?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too you young for this activity
}
\item During last week, easy/hard to:  Sit in a regular chair without holding on?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too you young for this activity
}
\item During last week, easy/hard to:  Get on and off a toilet or chair?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too you young for this activity
}
\item During last week, easy/hard to:  Get in and out of bed?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too you young for this activity
}
\item During last week, easy/hard to:  Turn door knobs?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too you young for this activity
}
\item During last week, easy/hard to:  Bend over from a standing position and pick up something off the floor?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all | 5, Too you young for this activity
}
\item How often need help from another person for sitting and standing?
\itemize{
\item 1, Never | 2, Sometimes | 3, About half the time | 4, Often | 5, All the time
}
\item How often use assistive devices for sitting and standing?
\itemize{
\item 1, Never | 2, Sometimes | 3, About half the time | 4, Often | 5, All the time
}
\item Participate in recreational outdoor activities with other children the same age?
\itemize{
\item 1, Yes easily | 2, Yes but a little hard | 3, Yes but very hard | 4, No
}
\item Was activity limited by: Pain?
\itemize{
\item 1, Yes
}
\item Was activity limited by: General Health?
\itemize{
\item 1, Yes
}
\item Was activity limited by: Doctor or parent instructions?
\itemize{
\item 1, Yes
}
\item Was activity limited by: Fear the other kids won’t like him/her?
\itemize{
\item 1, Yes
}
\item Was activity limited by: Dislike of recreational outdoor activities?
\itemize{
\item 1, Yes
}
\item Was activity limited by: Too young?
\itemize{
\item 1, Yes
}
\item Was activity limited by: Activity not in season?
\itemize{
\item 1, Yes
}
\item Participate in pickup games or sports with other children the same age?
\itemize{
\item 1, Yes easily | 2, Yes but a little hard | 3, Yes but very hard | 4, No
}
\item Was activity limited by: Pain?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  General Health?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Doctor or parent instructions?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Fear the other kids won’t like him/her?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Dislike of pickup games or sports?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Too young?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Activity not in season?
\itemize{
\item 1, Yes
}
\item Participate in competitive level sports with other children the same age?
\itemize{
\item 1, Yes, easily | 2, Yes, but a little hard | 3, Yes, but very hard | 4, No
}
\item Was activity limited by:  Pain?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  General Health?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Doctor or parent instructions?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Fear the other kids won’t like him/her?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Dislike of competitive level sports?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Too young?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Activity not in season?
\itemize{
\item 1, Yes
}
\item How often in last week did child get together and do things with friends?
\itemize{
\item 1, Often | 2, Sometimes | 3, Never or rarely
}
\item Was activity limited by:  Pain?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  General Health?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Doctor or parent instructions?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Fear the other kids won’t like him/her?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Friends not around?
\itemize{
\item 1, Yes
}
\item How often in last week did child participate in gym/recess?
\itemize{
\item 1, Often | 2, Sometimes | 3, Never or rarely | 4, No gym or recess
}
\item Was activity limited by:  Pain?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  General Health?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Doctor or parent instructions?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Fear the other kids won’t like him/her?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Dislike of gym/recess?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  School not in session?
\itemize{
\item 1, Yes
}
\item Was activity limited by:  Does not attend school?
\itemize{
\item 1, Yes
}
\item Is it easy or hard to make friends with children own age?
\itemize{
\item 1, Usually easy | 2, Sometimes easy | 3, Sometimes hard | 4, Usually hard
}
\item How much pain during the last week?
\itemize{
\item 1, None | 2, Very mild | 3, Mild | 4, Moderate | 5, Severe | 6, Very severe
}
\item During last week, how much did pain interfere with normal activities
\itemize{
\item 1, Not at all | 2, A little bit | 3, Moderately | 4, Quite a bit | 5, Extremely
}
\item Expectation of Treatment: To have pain relief.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item Expectation of Treatment: To look better.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item Expectation of Treatment: To feel better about himself/herself.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item Expectation of Treatment: To sleep more comfortably.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item Expectation of Treatment: To be able to do activities at home.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item Expectation of Treatment: To be able to do more at school.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item Expectation of Treatment: To be able to do more play or recreational activities.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item Expectation of Treatment: To be able to do more sports.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item Expectation of Treatment: To be free from pain or disability as an adult.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item If child had to spend the rest of his/her life with his/her bone and muscle condition as it is right now, how would you feel about it?
\itemize{
\item 1, Very satisfied | 2, Somewhat satisfied | 3, Neutral | 4, Somewhat dissatisfied | 5, Very dissatisfied
}
}
}
\usage{
podci_ado_prnt
}
\description{
Example Adolescent Parent-Reported Outcome Data Collection Instrument (PODCI) Data
}
\keyword{datasets}
