% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{podci_ado_self}
\alias{podci_ado_self}
\title{PODCI Adolescent Self-Report Example Data}
\format{
A 10 x 83 \link[dplyr:reexports]{dplyr::tibble}
\enumerate{
\item During the last week, was it easy or hard for you to: Lift heavy books?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week, was it easy or hard for you to: Pour a half gallon milk?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week, was it easy or hard for you to: Open a jar that has been opened before?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week, was it easy or hard for you to: Usa a fork and spoon?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week, was it easy or hard for you to: Comb your hair?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week, was it easy or hard for you to: Button buttons?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week, was it easy or hard for you to: Put on your coat?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week, was it easy or hard for you to: Write with a pencil?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week, was it easy or hard for you to: On average over the last 12 months, how often did you miss school (camp, etc) because of your health?
\itemize{
\item 1, Rarely | 2, Once a month | 3, Two or three times a month | 4, Once a week | 5, More than once a week | 6, Do not attend school, etc.
}
\item During the last week, how happy have you been with: How you look?
\itemize{
\item 1, Very happy | 2, Somewhat happy | 3, Not Sure | 4, Somewhat unhappy | 5, Very unhappy
}
\item During the last week, how happy have you been with: Your body?
\itemize{
\item 1, Very happy | 2, Somewhat happy | 3, Not Sure | 4, Somewhat unhappy | 5, Very unhappy
}
\item During the last week, how happy have you been with: What clothes or shoes you can wear?
\itemize{
\item 1, Very happy | 2, Somewhat happy | 3, Not Sure | 4, Somewhat unhappy | 5, Very unhappy
}
\item During the last week, how happy have you been with: Your ability to do the same things your friends do?
\itemize{
\item 1, Very happy | 2, Somewhat happy | 3, Not Sure | 4, Somewhat unhappy | 5, Very unhappy
}
\item During the last week, how happy have you been with: Your health in general?
\itemize{
\item 1, Very happy | 2, Somewhat happy | 3, Not Sure | 4, Somewhat unhappy | 5, Very unhappy
}
\item During the last week, how much of the time: Did you feel sick and tired?
\itemize{
\item 1, Most of the time | 2, Some of the time | 3, A little of the time | 4, None of the time
}
\item During the last week, how much of the time: Were you full of pep and energy?
\itemize{
\item 1, Most of the time | 2, Some of the time | 3, A little of the time | 4, None of the time
}
\item During the last week, how much of the time: Did pain or discomfort interfere with your activities?
\itemize{
\item 1, Most of the time | 2, Some of the time | 3, A little of the time | 4, None of the time
}
\item During the last week has it been easy or hard for you to: Run short distances?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week has it been easy or hard for you to: Bicycle or tricycle?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week has it been easy or hard for you to: Climb three flights of stairs?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week has it been easy or hard for you to: Climb one flight of stairs
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week has it been easy or hard for you to: Walk more than a mile?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week has it been easy or hard for you to: Walk three blocks?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week has it been easy or hard for you to: Walk one block?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week has it been easy or hard for you to: Get on and off a bus?
\itemize{
\item 1, Easy | 2, A little hard | 3, Very hard | 4, Can't do at all
}
\item During the last week has it been easy or hard for you to: How often do you need help form another person for walking and climbing?
\itemize{
\item 1, Never | 2, Sometimes | 3, About half the time | 4, Often | 5, All the time
}
\item During the last week has it been easy or hard for you to: How often do you use assistive devices (braces, crutches, wheelchair) for walking and climbing?
\itemize{
\item 1, Never | 2, Sometimes | 3, About half the time | 4, Often | 5, All the time
}
\item During the last week, has it been easy or hard for you to: Stand while washing your hands and face at a sink?
\itemize{
\item 1, Easy | 2, A little hard | 3, Vary hard | 4, Can't do at all
}
\item During the last week, has it been easy or hard for you to: Sit in a regular chair without holding on?
\itemize{
\item 1, Easy | 2, A little hard | 3, Vary hard | 4, Can't do at all
}
\item During the last week, has it been easy or hard for you to: Get on and off a toilet or chair?
\itemize{
\item 1, Easy | 2, A little hard | 3, Vary hard | 4, Can't do at all
}
\item During the last week, has it been easy or hard for you to: Get in and out of bed?
\itemize{
\item 1, Easy | 2, A little hard | 3, Vary hard | 4, Can't do at all
}
\item During the last week, has it been easy or hard for you to: Turn door knobs?
\itemize{
\item 1, Easy | 2, A little hard | 3, Vary hard | 4, Can't do at all
}
\item During the last week, has it been easy or hard for you to: Bend over from a  standing position and pick up something off the floor?
\itemize{
\item 1, Easy | 2, A little hard | 3, Vary hard | 4, Can't do at all
}
\item During the last week, has it been easy or hard for you to: How often do you need help from another person for sitting and standing?
\itemize{
\item 1, Never | 2, Sometimes | 3, About half the time | 4, Often | 5, All the time
}
\item During the last week, has it been easy or hard for you to: How often do you use assistive devices (braces, crutches, wheelchair) for sitting and standing?
\itemize{
\item 1, Never | 2, Sometimes | 3, About half the time | 4, Often | 5, All the time
}
\item During the last week, has it been easy or hard for you to: Can you participate in recreational outdoor activities with otter kids the same age?
\itemize{
\item 1, Yes easily | 2, Yes but a little hard | 3, yes but very hard | 4, No
}
\item During the last week, has it been easy or hard for you to: Can you participate in pickup games or sports with other kids the same age?
\itemize{
\item 1, Yes easily | 2, Yes but a little hard | 3, Yes but very hard | 4, No
}
\item During the last week, has it been easy or hard for you to: Can you participate in competitive level sports with other kids the same age?
\itemize{
\item 1, Yes easily | 2, Yes but a little hard | 3, Yes but very hard | 4, No
}
\item During the last week, has it been easy or hard for you to: How often in the last week did you get together and do things with friends?
\itemize{
\item 1, Often | 2, Sometimes | 3, Never or rarely
}
\item During the last week, has it been easy or hard for you to: How often in the last week did you participate in gym/recess?
\itemize{
\item 1, Often | 2, Sometimes | 3, Never or rarely | 4, No gym or recess
}
\item During the last week, has it been easy or hard for you to: Is it easy for you to make friends your own age?
\itemize{
\item 1, Usually easy | 2, Sometimes easy | 3, Sometimes hard | 4, Usually hard
}
\item During the last week, has it been easy or hard for you to: How much pain have you had during the last week?
\itemize{
\item 1, None | 2, Very mild | 3, Mild | 4, Moderate | 5, Severe | 6, Very Severe
}
\item During the last week, has it been easy or hard for you to: During the last week, how much did pain interfere with your normal activities?
\itemize{
\item 1, Not at all | 2, A little bit | 3, Moderately | 4, Quite a bit | 5, Extremely
}
\item As a result of my treatment, I expect: To have pain relief.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item As a result of my treatment, I expect: To look better.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item As a result of my treatment, I expect: To feel better about myself.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item As a result of my treatment, I expect: To sleep more comfortably.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item As a result of my treatment, I expect: To be able to do activities at home.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item As a result of my treatment, I expect: To be able to do more activities at school.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item As a result of my treatment, I expect: To be able to do more play or recreational activities (biking, walking, doing things with friends)
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item As a result of my treatment, I expect: To be able to do more sports.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item As a result of my treatment, I expect: To be free from pain or disability as an adult.
\itemize{
\item 1, Definitely yes | 2, Probably yes | 3, Not sure | 4, Probably not | 5, Definitely not
}
\item As a result of my treatment, I expect: If you had to spend the rest of your life with your bone and muscle condition as it is right now, how would you fee about it?
\itemize{
\item 1, Very satisfied | 2, Somewhat satisfied | 3, Neutral | 4, Somewhat dissatisfied | 5, Very dissatisfied
}
}
}
\usage{
podci_ado_self
}
\description{
Example Adolescent Self-Reported Outcome Data Collection Instrument (PODCI) Data
}
\keyword{datasets}
