% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\name{podci_items}
\alias{podci_items}
\title{PODCI Scale Item Numbers}
\usage{
podci_items(
  scale = c("upex", "tran", "sprt", "pain", "hapy"),
  type = c("both", "primary", "secondary")
)
}
\arguments{
\item{scale}{requested scale}

\item{type}{the type of item (ignored if not sport scale):
\itemize{
\item primary are items from which the scale score is directly computed
\item secondary are items for which the scoring of primary items depends
\item both gives both primary and secondary item numbers
}}
}
\value{
a vector of scale items
}
\description{
PODCI Scale Item Numbers
}
\examples{
## upper extremity item numbers
podci_items("upex", "both")

## sport primary and secondary item numbers
podci_items("sprt", "both")
}
