% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{podci_norms}
\alias{podci_norms}
\title{PODCI Normative Values}
\usage{
podci_norms(
  scale = c("upex", "tran", "sprt", "pain", "hapy", "glob"),
  patient = c("ped", "ado"),
  reporter = c("prnt", "self"),
  stat = c("m", "s")
)
}
\arguments{
\item{scale}{requested scale}

\item{patient}{subject of the assessment (pediatric vs adolescent)}

\item{reporter}{person providing the responses (parent vs self)}

\item{stat}{normative mean or standard deviation}
}
\value{
the requested normative value
}
\description{
PODCI Normative Values
}
\examples{
## adolescent happiness parent-report normative mean
podci_norms("hapy", "ado")

## pediatric sports normative standard deviation
podci_norms("sprt", stat = "s")

## adolescent pain self-report normative standard deviation
podci_norms("pain", "ado", "self", "s")

}
