% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pain.R
\name{podci_pain}
\alias{podci_pain}
\alias{podci_pain_raw_ped_prnt}
\alias{podci_pain_mean_ped_prnt}
\alias{podci_pain_stnd_ped_prnt}
\alias{podci_pain_norm_ped_prnt}
\alias{podci_pain_raw_ado_prnt}
\alias{podci_pain_mean_ado_prnt}
\alias{podci_pain_stnd_ado_prnt}
\alias{podci_pain_norm_ado_prnt}
\alias{podci_pain_raw_ado_self}
\alias{podci_pain_mean_ado_self}
\alias{podci_pain_stnd_ado_self}
\alias{podci_pain_norm_ado_self}
\title{PODCI Pain Scores}
\usage{
podci_pain(
  data,
  ...,
  score = c("raw", "mean", "stnd", "norm"),
  patient = c("ped", "ado"),
  reporter = c("prnt", "self"),
  norm_m,
  norm_s
)

podci_pain_raw_ped_prnt(data, ...)

podci_pain_mean_ped_prnt(data, ...)

podci_pain_stnd_ped_prnt(data, ...)

podci_pain_norm_ped_prnt(data, ...)

podci_pain_raw_ado_prnt(data, ...)

podci_pain_mean_ado_prnt(data, ...)

podci_pain_stnd_ado_prnt(data, ...)

podci_pain_norm_ado_prnt(data, ...)

podci_pain_raw_ado_self(data, ...)

podci_pain_mean_ado_self(data, ...)

podci_pain_stnd_ado_self(data, ...)

podci_pain_norm_ado_self(data, ...)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} containing the PODCI pain item responses
to be scored}

\item{...}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> columns of the pain items in
order (See Note).}

\item{score}{requested scale class
\itemize{
\item raw \verb{[3, 15]}
\item mean \verb{[1, 5]}
\item standard \verb{[0, 100]}
\item normative
\itemize{
\item pediatric \verb{[-17, 56]}
\item adolescent: \verb{[-1, 57]}
}
}}

\item{patient}{subject of the assessment (pediatric vs adolescent)}

\item{reporter}{person providing the responses (parent vs self)}

\item{norm_m}{mean value to use when computing normative scores
(See \link{podci_norms})}

\item{norm_s}{standard deviation value to use when computing normative scores
(See \link{podci_norms})}
}
\value{
data augmented with the requested score column
}
\description{
PODCI Pain Scores
}
\section{Functions}{
\itemize{
\item \code{podci_pain_raw_ped_prnt()}: Pain Raw Pediatric Parent Score

\item \code{podci_pain_mean_ped_prnt()}: Pain Mean Pediatric Parent Score

\item \code{podci_pain_stnd_ped_prnt()}: Pain Standard Pediatric Parent Score

\item \code{podci_pain_norm_ped_prnt()}: Pain Normal Pediatric Parent Score

\item \code{podci_pain_raw_ado_prnt()}: Pain Raw Adolescent Parent Score

\item \code{podci_pain_mean_ado_prnt()}: Pain Mean Adolescent Parent Score

\item \code{podci_pain_stnd_ado_prnt()}: Pain Standard Adolescent Parent Score

\item \code{podci_pain_norm_ado_prnt()}: Pain Normal Adolescent Parent Score

\item \code{podci_pain_raw_ado_self()}: Pain Raw Adolescent Self Score

\item \code{podci_pain_mean_ado_self()}: Pain Mean Adolescent Self Score

\item \code{podci_pain_stnd_ado_self()}: Pain Standard Adolescent Self Score

\item \code{podci_pain_norm_ado_self()}: Pain Normal Adolescent Self Score

}}
\note{
Pediatric
\enumerate{
\item Q17	Did pain or discomfort interfere with your child’s activities?
\item Q75	How much pain has your child had during the last week?
\item Q76	During the last week,  how much did pain interfere with your child’s
normal activities (including at home,  outside of the home,  and at school)?
}

A minimum of 2 items must have valid answers to score this scale.

Adolescent (Parent-Report)
\enumerate{
\item Q17	Did pain or discomfort interfere with your child’s activities?
\item Q75	How much pain has your child had during the last week?
\item Q76	During the last week,  how much did pain interfere with your child’s
normal activities (including at home,  outside of the home,  and at school)?
}

A minimum of 2 items must have valid answers to score this scale.

Adolescent (Self-Report)
\enumerate{
\item Q17 Did pain or discomfort interfere with your activities?
\item Q72 How much pain have you had during the last week?
\item Q73 During the last week,  how much did pain interfere with your normal
activities (including at home,  outside of the home,  and at school)?
}

A minimum of 2 items must have valid answers to score this scale.
}
\examples{
podci_pain_raw_ped_prnt(podci_ped_prnt, podci_items("pain", "ped", "prnt"))

podci_pain_mean_ped_prnt(podci_ped_prnt, podci_items("pain", "ped", "prnt"))

podci_pain_stnd_ped_prnt(podci_ped_prnt, podci_items("pain", "ped", "prnt"))

podci_pain_norm_ped_prnt(podci_ped_prnt, podci_items("pain", "ped", "prnt"))

podci_pain_raw_ado_prnt(podci_ado_prnt, podci_items("pain", "ado", "prnt"))

podci_pain_mean_ado_prnt(podci_ado_prnt, podci_items("pain", "ado", "prnt"))

podci_pain_stnd_ado_prnt(podci_ado_prnt, podci_items("pain", "ado", "prnt"))

podci_pain_norm_ado_prnt(podci_ado_prnt, podci_items("pain", "ado", "prnt"))

podci_pain_raw_ado_self(podci_ado_self, podci_items("pain", "ado", "self"))

podci_pain_mean_ado_self(podci_ado_self, podci_items("pain", "ado", "self"))

podci_pain_stnd_ado_self(podci_ado_self, podci_items("pain", "ado", "self"))

podci_pain_norm_ado_self(podci_ado_self, podci_items("pain", "ado", "self"))

}
