% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{round_podci_output_for_tests}
\alias{round_podci_output_for_tests}
\title{Round PODCI Output for Tests}
\usage{
round_podci_output_for_tests(data, digits = 2)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} of test data}

\item{digits}{integer indicating the number of decimal places
    (\code{round}) or significant digits (\code{signif}) to be used.
    For \code{round}, negative values are allowed (see \sQuote{Details}).}
}
\value{
a \link[dplyr:reexports]{dplyr::tibble} of rounded PODCI outputs
}
\description{
Round PODCI Output for Tests
}
\seealso{
\code{\link[=round]{round()}}
}
