% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glob.R
\name{podci_glob}
\alias{podci_glob}
\alias{podci_glob_stnd_ped}
\alias{podci_glob_norm_ped}
\alias{podci_glob_stnd_ado}
\alias{podci_glob_norm_ado}
\title{PODCI Global Scores}
\usage{
podci_glob(
  data,
  upex_stnd,
  tran_stnd,
  sprt_stnd,
  pain_stnd,
  score = c("stnd", "norm"),
  norm_m,
  norm_s
)

podci_glob_stnd_ped(data, upex_stnd, tran_stnd, sprt_stnd, pain_stnd)

podci_glob_norm_ped(data, upex_stnd, tran_stnd, sprt_stnd, pain_stnd)

podci_glob_stnd_ado(data, upex_stnd, tran_stnd, sprt_stnd, pain_stnd)

podci_glob_norm_ado(data, upex_stnd, tran_stnd, sprt_stnd, pain_stnd)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} containing the PODCI happy item responses
to be scored}

\item{upex_stnd}{upper extremity standard score}

\item{tran_stnd}{transfer standard score}

\item{sprt_stnd}{sport standard score}

\item{pain_stnd}{pain standard score}

\item{score}{requested scale class:
\itemize{
\item standard \verb{[0, 100]}
\item normative
\itemize{
\item pediatric \verb{[-77, 58]}
\item adolescent \verb{[-78, 57]}
}
}}

\item{norm_m}{mean value to use when computing normative scores
(See \link{podci_norms})}

\item{norm_s}{standard deviation value to use when computing normative scores
(See \link{podci_norms})}
}
\value{
data augmented with the requested score column
}
\description{
PODCI Global Scores
}
\section{Functions}{
\itemize{
\item \code{podci_glob_stnd_ped()}: PODCI Global Standard Pediatric Score

\item \code{podci_glob_norm_ped()}: PODCI Global Normal Pediatric Score

\item \code{podci_glob_stnd_ado()}: PODCI Global Standard Adolescent Score

\item \code{podci_glob_norm_ado()}: PODCI Global Normal Adolescent Score

}}
\examples{
## podci_glob_stnd_ped()
podci \%>\%
  podci_upex_stnd_ped(podci_items("upex")) \%>\%
  podci_tran_stnd_ped(podci_items("tran")) \%>\%
  podci_sprt_stnd_ped(podci_items("sprt")) \%>\%
  podci_pain_stnd_ped(podci_items("pain")) \%>\%
  podci_glob_stnd_ped(
    podci_upex_stnd_ped,
    podci_tran_stnd_ped,
    podci_sprt_stnd_ped,
    podci_pain_stnd_ped
  )

## podci_glob_norm_ped()
podci \%>\%
  podci_upex_stnd_ped(podci_items("upex")) \%>\%
  podci_tran_stnd_ped(podci_items("tran")) \%>\%
  podci_sprt_stnd_ped(podci_items("sprt")) \%>\%
  podci_pain_stnd_ped(podci_items("pain")) \%>\%
  podci_glob_norm_ped(
    podci_upex_stnd_ped,
    podci_tran_stnd_ped,
    podci_sprt_stnd_ped,
    podci_pain_stnd_ped
  )

## podci_glob_stnd_ado()
podci \%>\%
  podci_upex_stnd_ado(podci_items("upex")) \%>\%
  podci_tran_stnd_ado(podci_items("tran")) \%>\%
  podci_sprt_stnd_ado(podci_items("sprt")) \%>\%
  podci_pain_stnd_ado(podci_items("pain")) \%>\%
  podci_glob_stnd_ado(
    podci_upex_stnd_ado,
    podci_tran_stnd_ado,
    podci_sprt_stnd_ado,
    podci_pain_stnd_ado
  )

## podci_glob_norm_ado()
podci \%>\%
  podci_upex_stnd_ado(podci_items("upex")) \%>\%
  podci_tran_stnd_ado(podci_items("tran")) \%>\%
  podci_sprt_stnd_ado(podci_items("sprt")) \%>\%
  podci_pain_stnd_ado(podci_items("pain")) \%>\%
  podci_glob_norm_ado(
    podci_upex_stnd_ado,
    podci_tran_stnd_ado,
    podci_sprt_stnd_ado,
    podci_pain_stnd_ado
  )

}
