% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapy.R
\name{podci_hapy}
\alias{podci_hapy}
\alias{podci_hapy_raw_ped}
\alias{podci_hapy_mean_ped}
\alias{podci_hapy_stnd_ped}
\alias{podci_hapy_norm_ped}
\alias{podci_hapy_raw_ado}
\alias{podci_hapy_mean_ado}
\alias{podci_hapy_stnd_ado}
\alias{podci_hapy_norm_ado}
\title{PODCI Happy Scores}
\usage{
podci_hapy(data, ..., score = c("raw", "mean", "stnd", "norm"), norm_m, norm_s)

podci_hapy_raw_ped(data, ...)

podci_hapy_mean_ped(data, ...)

podci_hapy_stnd_ped(data, ...)

podci_hapy_norm_ped(data, ...)

podci_hapy_raw_ado(data, ...)

podci_hapy_mean_ado(data, ...)

podci_hapy_stnd_ado(data, ...)

podci_hapy_norm_ado(data, ...)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} containing the PODCI happy item responses
to be scored}

\item{...}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> columns of the happy items in
order (See Note).}

\item{score}{requested scale class:
\itemize{
\item raw \verb{[5, 25]}
\item mean \verb{[1, 5]}
\item standard \verb{[0, 100]}
\item normative
\itemize{
\item pediatric \verb{[-14, 57]}
\item adolescent \verb{[4, 60]}
}
}}

\item{norm_m}{mean value to use when computing normative scores}

\item{norm_s}{standard deviation value to use when computing normative scores}
}
\value{
data augmented with the requested score column
}
\description{
PODCI Happy Scores
}
\section{Functions}{
\itemize{
\item \code{podci_hapy_raw_ped()}: PODCI Happy Mean Pediatric Score

\item \code{podci_hapy_mean_ped()}: PODCI Happy Mean Pediatric Score

\item \code{podci_hapy_stnd_ped()}: PODCI Happy Standard Pediatric Score

\item \code{podci_hapy_norm_ped()}: PODCI Happy Normal Pediatric Score

\item \code{podci_hapy_raw_ado()}: PODCI Happy Mean Adolescent Score

\item \code{podci_hapy_mean_ado()}: PODCI Happy Mean Adolescent Score

\item \code{podci_hapy_stnd_ado()}: PODCI Happy Standard Adolescent Score

\item \code{podci_hapy_norm_ado()}: PODCI Happy Normal Adolescent Score

}}
\note{
The scale names listed in \code{...} are expected to be in this order:
\enumerate{
\item Q10	How he/she looks?
\item Q11	His/her body?
\item Q12	What clothes or shoes he/she can wear?
\item Q13	His/her ability to do the same things his/her friends do?
\item Q14	His/her health in general?
}
}
\examples{
podci_hapy_raw_ped(podci, podci_items("hapy"))
podci_hapy_mean_ped(podci, podci_items("hapy"))
podci_hapy_stnd_ped(podci, podci_items("hapy"))
podci_hapy_norm_ped(podci, podci_items("hapy"))
podci_hapy_raw_ado(podci, podci_items("hapy"))
podci_hapy_mean_ado(podci, podci_items("hapy"))
podci_hapy_stnd_ado(podci, podci_items("hapy"))
podci_hapy_norm_ado(podci, podci_items("hapy"))
}
