% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\name{podci_items}
\alias{podci_items}
\title{PODCI Scale Item Numbers}
\usage{
podci_items(
  scale = c("upex", "tran", "sprt", "pain", "hapy", "glob"),
  patient = c("ped", "ado"),
  reporter = c("prnt", "self"),
  type = c("both", "primary", "secondary")
)
}
\arguments{
\item{scale}{requested scale}

\item{patient}{subject of the assessment (pediatric vs adolescent)}

\item{reporter}{person providing the responses (parent vs self)}

\item{type}{the type of item (ignored if not sport scale):
\itemize{
\item primary are items from which the scale score is directly computed
\item secondary are items for which the scoring of primary items depends
\item both gives both primary and secondary item numbers
}}
}
\value{
a vector of scale items
}
\description{
PODCI Scale Item Numbers
}
\examples{
## all scales have the same item numbers except:

### the sport scale items differ depending on reporter
### and has primary and secondary items

podci_items("sprt", reporter = "prnt", type = "both")

podci_items("sprt", reporter = "prnt", type = "primary")

podci_items("sprt", reporter = "prnt", type = "secondary")

podci_items("sprt", reporter = "self", type = "both")

podci_items("sprt", reporter = "self", type = "primary")

podci_items("sprt", reporter = "self", type = "secondary")

### the pain scale items differ depending on reporter

podci_items("pain", reporter = "prnt")

podci_items("pain", reporter = "self")

### the global scale items differ depending on reporter

podci_items("glob", reporter = "prnt")

podci_items("glob", reporter = "self")

}
