% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{podci_norms}
\alias{podci_norms}
\title{PODCI Normative Values}
\usage{
podci_norms(
  pop = c("ped", "ado"),
  scale = c("upex", "tran", "sprt", "pain", "hapy", "glob"),
  stat = c("m", "s")
)
}
\arguments{
\item{pop}{population}

\item{scale}{requested scale}

\item{stat}{normative mean or standard deviation}
}
\value{
the requested normative value
}
\description{
PODCI Normative Values
}
\examples{
## adolescent happiness normative mean
podci_norms("ado", "hapy", "m")

## pediatric sports normative standard deviation
podci_norms("ped", "sprt", "s")

}
