% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pain.R
\name{podci_pain}
\alias{podci_pain}
\alias{podci_pain_raw_ped}
\alias{podci_pain_mean_ped}
\alias{podci_pain_stnd_ped}
\alias{podci_pain_norm_ped}
\alias{podci_pain_raw_ado}
\alias{podci_pain_mean_ado}
\alias{podci_pain_stnd_ado}
\alias{podci_pain_norm_ado}
\title{PODCI Pain Scores}
\usage{
podci_pain(data, ..., score = c("raw", "mean", "stnd", "norm"), norm_m, norm_s)

podci_pain_raw_ped(data, ...)

podci_pain_mean_ped(data, ...)

podci_pain_stnd_ped(data, ...)

podci_pain_norm_ped(data, ...)

podci_pain_raw_ado(data, ...)

podci_pain_mean_ado(data, ...)

podci_pain_stnd_ado(data, ...)

podci_pain_norm_ado(data, ...)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} containing the PODCI pain item responses
to be scored}

\item{...}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> columns of the pain items in
order (See Note).}

\item{score}{requested scale class
\itemize{
\item raw \verb{[3, 15]}
\item mean \verb{[1, 5]}
\item standard \verb{[0, 100]}
\item normative
\itemize{
\item pediatric \verb{[-17, 56]}
\item adolescent: \verb{[-1, 57]}
}
}}

\item{norm_m}{mean value to use when computing normative scores}

\item{norm_s}{standard deviation value to use when computing normative scores}
}
\value{
data augmented with the requested score column
}
\description{
PODCI Pain Scores
}
\section{Functions}{
\itemize{
\item \code{podci_pain_raw_ped()}: PODCI Pain Raw Pediatric Score

\item \code{podci_pain_mean_ped()}: PODCI Pain Mean Pediatric Score

\item \code{podci_pain_stnd_ped()}: PODCI Pain Standard Pediatric Score

\item \code{podci_pain_norm_ped()}: PODCI Pain Normal Pediatric Score

\item \code{podci_pain_raw_ado()}: PODCI Pain Raw Adolescent Score

\item \code{podci_pain_mean_ado()}: PODCI Pain Mean Adolescent Score

\item \code{podci_pain_stnd_ado()}: PODCI Pain Standard Adolescent Score

\item \code{podci_pain_norm_ado()}: PODCI Pain Normal Adolescent Score

}}
\note{
The scale names listed in \code{...} are expected to be in this order:
\enumerate{
\item Q17	Did pain or discomfort interfere with your child’s activities?
\item Q75	How much pain has your child had during the last week?
\item Q76	During the last week,  how much did pain interfere with your child’s
normal activities (including at home,  outside of the home,  and at school)?
}

A minimum of 2 items must have valid answers to score this scale.
}
\examples{
podci_pain_raw_ped(podci, podci_items("pain"))
podci_pain_mean_ped(podci, podci_items("pain"))
podci_pain_stnd_ped(podci, podci_items("pain"))
podci_pain_norm_ped(podci, podci_items("pain"))
podci_pain_raw_ado(podci, podci_items("pain"))
podci_pain_mean_ped(podci, podci_items("pain"))
podci_pain_stnd_ped(podci, podci_items("pain"))
podci_pain_norm_ped(podci, podci_items("pain"))
}
