% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprt.R
\name{podci_sprt}
\alias{podci_sprt}
\alias{podci_sprt_raw_ped}
\alias{podci_sprt_mean_ped}
\alias{podci_sprt_stnd_ped}
\alias{podci_sprt_norm_ped}
\alias{podci_sprt_raw_ado}
\alias{podci_sprt_mean_ado}
\alias{podci_sprt_stnd_ado}
\alias{podci_sprt_norm_ado}
\title{PODCI Sport Scores}
\usage{
podci_sprt(data, ..., score = c("raw", "mean", "stnd", "norm"), norm_m, norm_s)

podci_sprt_raw_ped(data, ...)

podci_sprt_mean_ped(data, ...)

podci_sprt_stnd_ped(data, ...)

podci_sprt_norm_ped(data, ...)

podci_sprt_raw_ado(data, ...)

podci_sprt_mean_ado(data, ...)

podci_sprt_stnd_ado(data, ...)

podci_sprt_norm_ado(data, ...)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} containing the PODCI sport item responses
to be scored}

\item{...}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> columns of the sport items in
order (See Note).}

\item{score}{requested scale class
\itemize{
\item raw \verb{[12, 48]}
\item mean \verb{[1, 4]}
\item standard \verb{[0, 100]}
\item normative
\itemize{
\item pediatric \verb{[-40, 57]}
\item adolscent: \verb{[-36, 56]}
}
}}

\item{norm_m}{mean value to use when computing normative scores}

\item{norm_s}{standard deviation value to use when computing normative scores}
}
\value{
data augmented with the requested score column
}
\description{
PODCI Sport Scores
}
\section{Functions}{
\itemize{
\item \code{podci_sprt_raw_ped()}: PODCI Sport Raw Pediatric Score

\item \code{podci_sprt_mean_ped()}: PODCI Sport Mean Pediatric Score

\item \code{podci_sprt_stnd_ped()}: PODCI Sport Standard Pediatric Score

\item \code{podci_sprt_norm_ped()}: PODCI Sport Normal Pediatric Score

\item \code{podci_sprt_raw_ado()}: PODCI Sport Raw Adolescent Score

\item \code{podci_sprt_mean_ado()}: PODCI Sport Mean Adolescent Score

\item \code{podci_sprt_stnd_ado()}: PODCI Sport Standard Adolescent Score

\item \code{podci_sprt_norm_ado()}: PODCI Sport Normal Adolescent Score

}}
\note{
The scale names listed in \code{...} are expected to be in this order:
\enumerate{
\item Q18	Run short distances?
\item Q19	Bicycle or tricycle?
\item Q20	Climb three flights of stairs?
\item Q22	Walk more than a mile?
\item Q23	Walk three blocks?
\item Q26	How often does your child need help from another person for walking
and climbing?
\item Q27	How often does your child use assistive devices (such as braces,
crutches, or wheelchair) for walking and climbing?
\item Q36	Can your child participate in recreational outdoor activities with
other children the same age?
\item Q42 Too young?
\item Q43 Activity not in season?
\item Q44	Can your child participate in pickup games or sports with other
children the same age?
\item Q50 Too young?
\item Q51 Activity not in season?
\item Q52	Can your child participate in competitive level sports with other
children the same age?
\item Q58 Too young?
\item Q59 Activity not in season?
\item Q60	How often in the last week did your child get together and do things
with friends?
\item Q65 Friends not around?
\item Q66	How often in the last week did your child participate in gym/recess?
\item Q72 School not in session?
\item Q73 Does not attend school?
}

Any item rated "5" (Too young for this activity) is considered missing and
is not added into the scale.

A minimum of 6 items must have valid answers to score this scale (including
those marked "too young" as missing).
}
\examples{
podci_sprt_raw_ped(podci, podci_items("sprt"))
podci_sprt_mean_ped(podci, podci_items("sprt"))
podci_sprt_stnd_ped(podci, podci_items("sprt"))
podci_sprt_norm_ped(podci, podci_items("sprt"))
podci_sprt_raw_ado(podci, podci_items("sprt"))
podci_sprt_mean_ado(podci, podci_items("sprt"))
podci_sprt_stnd_ado(podci, podci_items("sprt"))
podci_sprt_norm_ado(podci, podci_items("sprt"))
}
