% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tran.R
\name{podci_tran}
\alias{podci_tran}
\alias{podci_tran_raw_ped}
\alias{podci_tran_mean_ped}
\alias{podci_tran_stnd_ped}
\alias{podci_tran_norm_ped}
\alias{podci_tran_raw_ado}
\alias{podci_tran_mean_ado}
\alias{podci_tran_stnd_ado}
\alias{podci_tran_norm_ado}
\title{PODCI Transfer Scores}
\usage{
podci_tran(data, ..., score = c("raw", "mean", "stnd", "norm"), norm_m, norm_s)

podci_tran_raw_ped(data, ...)

podci_tran_mean_ped(data, ...)

podci_tran_stnd_ped(data, ...)

podci_tran_norm_ped(data, ...)

podci_tran_raw_ado(data, ...)

podci_tran_mean_ado(data, ...)

podci_tran_stnd_ado(data, ...)

podci_tran_norm_ado(data, ...)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} containing the PODCI transfer item responses
to be scored}

\item{...}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> columns of the transfer items in
order (See Note).}

\item{score}{requested scale class
\itemize{
\item raw \verb{[11, 44]}
\item mean \verb{[1, 4]}
\item standard \verb{[0, 100]}
\item normative
\itemize{
\item pediatric \verb{[-123, 53]}
\item adolescent: \verb{[-164, 52]}
}
}}

\item{norm_m}{mean value to use when computing normative scores
(See \link{podci_norms})}

\item{norm_s}{standard deviation value to use when computing normative scores
(See \link{podci_norms})}
}
\value{
data augmented with the requested score column
}
\description{
PODCI Transfer Scores
}
\section{Functions}{
\itemize{
\item \code{podci_tran_raw_ped()}: PODCI Transfer Raw Pediatric Score

\item \code{podci_tran_mean_ped()}: PODCI Transfer Mean Pediatric Score

\item \code{podci_tran_stnd_ped()}: PODCI Transfer Standard Pediatric Score

\item \code{podci_tran_norm_ped()}: PODCI Transfer Normal Pediatric Score

\item \code{podci_tran_raw_ado()}: PODCI Transfer Raw Adolescent Score

\item \code{podci_tran_mean_ado()}: PODCI Transfer Mean Adolescent Score

\item \code{podci_tran_stnd_ado()}: PODCI Transfer Standard Adolescent Score

\item \code{podci_tran_norm_ado()}: PODCI Transfer Normal Adolescent Score

}}
\note{
The scale names listed in \code{...} are expected to be in this order:
\enumerate{
\item Q7	Put on his/her coat?
\item Q21	Climb one flight of stairs?
\item Q24	Walk one block?
\item Q25	Get on and off a bus?
\item Q28	Stand while washing his/her hands and face at a sink?
\item Q29	Sit in a regular chair without holding on?
\item Q30	Get on and off a toilet or chair?
\item Q31	Get in and out of bed?
\item Q33	Bend over from a standing position and pick up something off the
floor?
\item Q34	How often does your child need help from another person for sitting
and standing?
\item Q35	How often does your child use assistive devices (such as braces,
crutches, or wheelchair) for sitting and standing?
}

Any item rated "5" (Too young for this activity) is considered missing and
is not added into the scale.

A minimum of 7 items must have valid answers to score this scale (including
those marked "too young" as missing).
}
\examples{
podci_tran_raw_ped(podci, podci_items("tran"))

podci_tran_mean_ped(podci, podci_items("tran"))

podci_tran_stnd_ped(podci, podci_items("tran"))

podci_tran_norm_ped(podci, podci_items("tran"))

podci_tran_raw_ado(podci, podci_items("tran"))

podci_tran_mean_ado(podci, podci_items("tran"))

podci_tran_stnd_ado(podci, podci_items("tran"))

podci_tran_norm_ado(podci, podci_items("tran"))

}
