% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upex.R
\name{podci_upex}
\alias{podci_upex}
\alias{podci_upex_raw_ped}
\alias{podci_upex_mean_ped}
\alias{podci_upex_stnd_ped}
\alias{podci_upex_norm_ped}
\alias{podci_upex_raw_ado}
\alias{podci_upex_mean_ado}
\alias{podci_upex_stnd_ado}
\alias{podci_upex_norm_ado}
\title{PODCI Upper Extremity Scores}
\usage{
podci_upex(data, ..., score = c("raw", "mean", "stnd", "norm"), norm_m, norm_s)

podci_upex_raw_ped(data, ...)

podci_upex_mean_ped(data, ...)

podci_upex_stnd_ped(data, ...)

podci_upex_norm_ped(data, ...)

podci_upex_raw_ado(data, ...)

podci_upex_mean_ado(data, ...)

podci_upex_stnd_ado(data, ...)

podci_upex_norm_ado(data, ...)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} containing the PODCI upper extremity item
responses to be scored}

\item{...}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> columns of the upper extremity items
in order (See Note).}

\item{score}{requested scale class
\itemize{
\item raw \verb{[8, 32]}
\item mean \verb{[1, 4]}
\item standard \verb{[0, 100]}
\item normal
\itemize{
\item pediatric \verb{[-30, 57]}
\item adolescent \verb{[-146, 53]}
}
}}

\item{norm_m}{mean value to use when computing normative scores}

\item{norm_s}{standard deviation value to use when computing normative scores}
}
\value{
data augmented with the requested score column
}
\description{
PODCI Upper Extremity Scores
}
\section{Functions}{
\itemize{
\item \code{podci_upex_raw_ped()}: PODCI Upper Extremity Raw Pediatric Score

\item \code{podci_upex_mean_ped()}: PODCI Upper Extremity Mean Pediatric Score

\item \code{podci_upex_stnd_ped()}: PODCI Upper Extremity Standard Pediatric Score

\item \code{podci_upex_norm_ped()}: PODCI Upper Extremity Normal Pediatric Score

\item \code{podci_upex_raw_ado()}: PODCI Upper Extremity Raw Adolescent Score

\item \code{podci_upex_mean_ado()}: PODCI Upper Extremity Mean Adolescent Score

\item \code{podci_upex_stnd_ado()}: PODCI Upper Extremity Standard Adolescent Score

\item \code{podci_upex_norm_ado()}: PODCI Upper Extremity Normal Adolescent Score

}}
\note{
The scale names listed in \code{...} are expected to be in this order:
\enumerate{
\item Q1	Lift heavy books?
\item Q2	Pour a half gallon of milk?
\item Q3	Open a jar that has been opened before?
\item Q4	Use a fork and spoon?
\item Q5	Comb his/her hair?
\item Q6	Button buttons?
\item Q8	Write with a pencil?
\item Q32 Turn door knobs?
}

Any item rated "5" (Too young for this activity) is considered missing and
is not added into the scale.

A minimum of 4 items must have valid answers to score this scale (including
those marked "too young" as missing).
}
\examples{
podci_upex_raw_ped(podci, podci_items("upex"))
podci_upex_mean_ped(podci, podci_items("upex"))
podci_upex_stnd_ped(podci, podci_items("upex"))
podci_upex_norm_ped(podci, podci_items("upex"))
podci_upex_raw_ado(podci, podci_items("upex"))
podci_upex_mean_ado(podci, podci_items("upex"))
podci_upex_stnd_ado(podci, podci_items("upex"))
podci_upex_norm_ado(podci, podci_items("upex"))
}
